Move tree manipulation code about
