+++ Cargo.toml
@@ -72,6 +72,7 @@ but-rebase = { path = "crates/but-rebase" }
 but-core = { path = "crates/but-core" }
 but-workspace = { path = "crates/but-workspace" }
 but-hunk-dependency = { path = "crates/but-hunk-dependency" }
+but-status = { path = "crates/but-status" }
 
 [profile.release]
 codegen-units = 1 # Compile crates one after another so the compiler can optimize better
