+++ crates/but-status/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "but-status"
+version = "0.0.0"
+edition = "2024"
+authors = ["GitButler <gitbutler@gitbutler.com>"]
+publish = false
+
+[dependencies]
+gix.workspace = true
+anyhow.workspace = true
+bstr.workspace = true
+serde_json = "1.0.140"
+serde.workspace = true
